online node mode. First version.

+++ node/MwcNode.cpp
@@ -22,6 +22,7 @@
 #include "../core/Config.h"
 #include "../control/messagebox.h"
 #include "../core/Notification.h"
+#include "../core/Config.h"
 #include "../util/Log.h"
 #include <QNetworkAccessManager>
 #include <QNetworkReply>
@@ -31,6 +32,7 @@
 #include <QJsonArray>
 #include "MwcNodeConfig.h"
 #include <QTimer>
+#include <QCoreApplication>
 
 namespace node {
 
@@ -95,16 +97,22 @@ void MwcNode::start( const QString & network ) {
 void MwcNode::stop() {
     qDebug() << "MwcNode::stop ...";
 
+    QCoreApplication::processEvents();
+
     nodeProcDisconnect();
 
     if (nodeProcess) {
-        nodeProcess->kill();
-        if (!util::processWaitForFinished( nodeProcess, 30000, "mwc-node")) {
-            nodeProcess->terminate();
+
+        if (nodeProcess->state() == QProcess::Running) {
+            qDebug() << "killing mwc-node";
+            nodeProcess->kill();
+            if (!util::processWaitForFinished( nodeProcess, 30000, "mwc-node")) {
+                nodeProcess->terminate();
+            }
         }
         qDebug() << "mwc-node is exited";
 
-        notify::appendNotificationMessage( notify::MESSAGE_LEVEL::INFO, "Embedded mwc-node is stopped" );
+        notify::appendNotificationMessage( notify::MESSAGE_LEVEL::INFO, "Embedded mwc-node is stopped" + (restartCounter>0 ? ". Attempt " + QString::number(restartCounter+1) : "") );
 
         nodeProcess->deleteLater();
         nodeProcess = nullptr;
@@ -115,6 +123,7 @@ void MwcNode::stop() {
         nodeOutputParser = nullptr;
     }
 
+    QCoreApplication::processEvents();
 }
 
 // pass - provide password through env variable. If pass empty - nothing will be done
@@ -134,6 +143,8 @@ QProcess * MwcNode::initNodeProcess( QString network ) {
 
     // Working dir must match config file
     process->setWorkingDirectory(nodeWorkDir);
+
+    nodeProcConnect(process);
 
     QStringList params;
     if (network.toLower().startsWith("floo"))
@@ -163,8 +174,6 @@ QProcess * MwcNode::initNodeProcess( QString network ) {
                 return nullptr;
         }
     }
-
-    nodeProcConnect(process);
 
     return process;
 }
@@ -225,12 +234,13 @@ void MwcNode::nodeProcessFinished(int exitCode, QProcess::ExitStatus exitStatus)
 
     qDebug() << "mwc-node is exiting with exit code " << exitCode << ", exitStatus=" << exitStatus;
 
-    if (nodeProcess) {
+    stop();
+
+    Q_ASSERT(nodeProcess==nullptr);
+/*    if (nodeProcess) {
         nodeProcess->deleteLater();
         nodeProcess = nullptr;
-    }
-
-    stop();
+    }*/
 
     if (restartCounter<2) {
         MwcNodeConfig mainnetConfig = getCurrentMwcNodeConfig( "Mainnet" );
@@ -245,8 +255,8 @@ void MwcNode::nodeProcessFinished(int exitCode, QProcess::ExitStatus exitStatus)
 
         restartCounter++;
 
-        // restart the node in 2 seconds
-        QTimer::singleShot( 1000*2, this, &MwcNode::onRestartNode );
+        // restart the node in 4 seconds
+        QTimer::singleShot( 1000*4, this, &MwcNode::onRestartNode );
     }
     else {
         reportNodeFatalError( "mwc-node process exited due some unexpected error. The exit code: " + QString::number(exitCode) + "\n\n"
@@ -771,15 +781,20 @@ void MwcNode::replyFinished(QNetworkReply* reply) {
 
 void MwcNode::reportNodeFatalError( QString message ) {
 
-    if ( control::MessageBox::RETURN_CODE::BTN2 == control::MessageBox::question(nullptr, "Embedded MWC-Node Error",
-            message + "\n\nIf Embedded mwc-node doesn't work for you, please switch to MWC Cloud node before exit", "Keep Embedded", "Switch to Cloud", false, true ) ) {
-
-        // Switching to the cloud node
-        wallet::MwcNodeConnection mwcNodeConnection = appContext->getNodeConnection( lastUsedNetwork );
-        mwcNodeConnection.setData( wallet::MwcNodeConnection::NODE_CONNECTION_TYPE::CLOUD );
-        appContext->updateMwcNodeConnection(lastUsedNetwork, mwcNodeConnection );
+    if ( config::isOnlineNode() ) {
+        control::MessageBox::message(nullptr, "Embedded MWC-Node Error", message);
     }
+    else
+    {
+        if ( control::MessageBox::RETURN_CODE::BTN2 == control::MessageBox::question(nullptr, "Embedded MWC-Node Error",
+                message + "\n\nIf Embedded mwc-node doesn't work for you, please switch to MWC Cloud node before exit", "Keep Embedded", "Switch to Cloud", false, true ) ) {
 
+            // Switching to the cloud node
+            wallet::MwcNodeConnection mwcNodeConnection = appContext->getNodeConnection( lastUsedNetwork );
+            mwcNodeConnection.setData( wallet::MwcNodeConnection::NODE_CONNECTION_TYPE::CLOUD );
+            appContext->updateMwcNodeConnection(lastUsedNetwork, mwcNodeConnection );
+        }
+    }
 
     mwc::closeApplication();
